# products/data-cloud/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Data-Cloud Platform API
  description: |
    REST API for the Ghatana Data-Cloud unified data platform.

    Data-Cloud is an independent, deployable, AI/ML-native product. It owns data management,
    event-cloud streaming, analytics, reporting, governance, plugin-driven extensibility, feature
    storage, model metadata, and execution persistence surfaces.

    Data-Cloud now exposes plugin-backed workflow execution endpoints for pipeline execution,
    execution status inspection, and execution log retrieval inside the standalone launcher.
    AEP still owns broader agentic orchestration, but Data-Cloud can execute and observe
    pipeline workflows through its runtime plugin layer.

    The Data-Cloud platform provides:
    - **Entity storage** — multi-tenant JSONB entity CRUD with pluggable storage backends
    - **Event log** — append-only event stream with push-based SSE subscriptions
    - **Execution metadata and runtime** — store and retrieve pipeline/checkpoint records and execute pipeline workflows through runtime plugins
    - **Agent memory plane** — persist, read, and search agent episodic/semantic/procedural memory
    - **Brain API** — salience scoring, attention management, workspace spotlight
    - **Learning API** — policy extraction, human-review queue
    - **Analytics API** — ad-hoc ClickHouse-backed analytics queries
    - **Server-Sent Events** — real-time push for events and learning updates
...

    ## WebSocket
...

...

...
servers:
...
components:
  parameters:
    tenantIdQuery:
      name: tenantId
      in: query
      description: |
        Tenant identifier. Required for all entity, event, execution-metadata, and memory operations.
        May alternatively be supplied as the `X-Tenant-ID` request header.
      schema:
        type: string
        example: "tenant-123"
...
    limitQuery:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 1000
    offsetQuery:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
  schemas:
...
    Alert:
      type: object
      required: [id, tenantId, title, description, severity, status, source, createdAt]
      properties:
        id:
          type: string
        tenantId:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, warning, info]
        status:
          type: string
          enum: [active, acknowledged, resolved]
        source:
          type: string
        createdAt:
          type: string
          format: date-time
        acknowledgedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
    AlertListResponse:
      type: object
      required: [tenantId, alerts, count, timestamp]
      properties:
        tenantId:
          type: string
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        count:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
...
    AlertRule:
      type: object
      required: [id, tenantId, name, enabled, severity, conditionType, metric, operator, threshold, duration, channels]
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        severity:
          type: string
          enum: [critical, warning, info]
        conditionType:
          type: string
          enum: [threshold, anomaly, pattern, absence]
        metric:
          type: string
        operator:
          type: string
          enum: [gt, lt, eq, gte, lte]
        threshold:
          type: number
        duration:
          type: integer
          format: int32
        channels:
          type: array
          items:
            type: string
            enum: [email, slack, webhook, pagerduty]
        recipients:
          type: array
          items:
            type: string
        webhookUrl:
          type: string
    AlertRuleListResponse:
      type: object
      required: [tenantId, rules, count, timestamp]
      properties:
        tenantId:
          type: string
        rules:
          type: array
          items:
            $ref: "#/components/schemas/AlertRule"
        count:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: integer
        timestamp:
          type: string
          format: date-time
...
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /api/v1/alerts:
    get:
      summary: List operational alerts
      operationId: apiV1AlertsGet
      tags: [alerts]
      parameters:
        - $ref: "#/components/parameters/tenantIdQuery"
        - $ref: "#/components/parameters/limitQuery"
        - $ref: "#/components/parameters/offsetQuery"
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, warning, info]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, acknowledged, resolved]
      responses:
        "200":
          description: Alert list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertListResponse"
  /api/v1/alerts/{alertId}/acknowledge:
    post:
      summary: Acknowledge an alert
      operationId: apiV1AlertsAlertIdAcknowledgePost
      tags: [alerts]
      parameters:
        - $ref: "#/components/parameters/tenantIdQuery"
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert acknowledged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/alerts/{alertId}/resolve:
    post:
      summary: Resolve an alert
      operationId: apiV1AlertsAlertIdResolvePost
      tags: [alerts]
      parameters:
        - $ref: "#/components/parameters/tenantIdQuery"
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert resolved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/alerts/rules:
    get:
      summary: List alert rules
      operationId: apiV1AlertsRulesGet
      tags: [alerts]
      parameters:
        - $ref: "#/components/parameters/tenantIdQuery"
      responses:
        "200":
          description: Alert rules.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRuleListResponse"
    post:
      summary: Create an alert rule
      operationId: apiV1AlertsRulesPost
      tags: [alerts]
      parameters:
        - $ref: "#/components/parameters/tenantIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertRule"
      responses:
        "201":
          description: Alert rule created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
...